Hiển thị sản phẩm theo danh mục

+++ LuxubuShop.Web/Controllers/AboutController.cs
@@ -10,7 +10,7 @@ namespace LuxubuShop.Web.Controllers
     public class AboutController : Controller
     {
         // GET: About
-        public ActionResult Index(int id = 1)
+        public ActionResult Index(long id = 1)
         {
             var about = new AboutDao().ViewDetail(id);
             return View(about);
